Save new authticket on refresh
`authorize_app()` saves the authticket to `static::$instance` while `refreshAuthTicket()` used `$this`. It looks like this is trying to follow the Singleton pattern, so those should be the same object. Running a program that makes many API calls (like importing a whole store) often runs over the expiration time of the original authticket, but every call after that seems to require a refresh, which should not be necessary. Saving the new authticket to the static instance ought to fix that problem, doing only one refresh every hour until the refresh token expires.